app/requirements.txt
langgraph==0.2.25
langchain==0.2.14
langchain-community==0.2.12
pandas==2.2.2
pyarrow==16.1.0
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.8.2
xlwt==1.3.0
xlrd==2.0.1
XlsxWriter==3.2.0
rapidfuzz==3.9.6
chromadb==0.5.4
requests==2.32.3
